TD-21: Agregando nuevos test de tipo regression y modificando el nombre del archivo test
